// src/types/inquiryData.ts
export default interface InquiryData {
    Kunnr: string;
    Erdat: string;
    Auart: string;
    Angdt: string;
    Bnddt: string;
    Vbeln: string;
    Posnr: string;
    Netwr: string;
    Waerk: string;
    Arktx: string;
    Posar: string;
    Vrkme: string;
    Kwmeng: string;
};